refactor some utility code to CameraUtils
